Move Spec workflow tools to release plan hierarchy
